feat: refactor IterableEmbedded components to utilize useEmbeddedView hook
