+++ src/embedded/types/IterableEmbeddedViewProps.ts
