+++ src/embedded/utils/getDefaultActionUrl.ts
@@ -1,4 +1,4 @@
-import type { IterableEmbeddedMessage } from '../../classes/IterableEmbeddedMessage';
+import type { IterableEmbeddedMessage } from '../classes/IterableEmbeddedMessage';
 
 export const getDefaultActionUrl = (message: IterableEmbeddedMessage) => {
   const defaultAction = message.elements?.defaultAction ?? null;
