+++ src/embedded/utils/getUrlFromButton.ts
@@ -1,4 +1,4 @@
-import type { IterableEmbeddedMessageElementsButton } from '../../classes/IterableEmbeddedMessageElementsButton';
+import type { IterableEmbeddedMessageElementsButton } from '../classes/IterableEmbeddedMessageElementsButton';
 
 export const getUrlFromButton = (
   button: IterableEmbeddedMessageElementsButton
